change tg_id from int to bigint

+++ src/db/models.py
@@ -38,7 +38,7 @@ class UsersOrm(Base):
     __tablename__ = "users"
 
     id: Mapped[intpk]
-    tg_id: Mapped[int] = mapped_column(unique=True)
+    tg_id: Mapped[BIGINT] = mapped_column(unique=True)
     language: Mapped[str]
     timezone: Mapped[str] = mapped_column(nullable=False)
     notifications: Mapped[bool]
